To make ExternalMechanism.java class RFC compliant
The SASL External mechanism is configured in this class (ExternalMechanism.java) of Apache qpid-jms.

We can notice that the initial response is configured in Line 28. It is always set to EMPTY (empty byte array defined here) and cannot be configured to a custom string that we can use as identity string. Hence this library is NOT RFC compliant

If we have to pass the authorization identity string to the server then we must configure the initial response of that client.